+++ src/components/data/properties/camera/CameraProperties.vue
@@ -3,9 +3,9 @@
 		<MxAccordionItem label="Lens" :item="{ value: 'lens' }">
 			<div class="p-1 flex flex-col gap-0.5 items-end">
 				<InputField label="Type" input-width="150px" class="mb-1">
-					<InputSelect v-model="cameraType" :items="TYPE_OPTIONS" />
+					<InputSelect v-model="cameraStore.viewportCameraType" :items="TYPE_OPTIONS" />
 				</InputField>
-				<template v-if="isPerspectiveCamera(threeStore.activeCamera)">
+				<template v-if="isPerspectiveCamera(cameraStore.activeCamera)">
 					<InputField
 						v-if="currentLensUnit === 'fov'"
 						label="Field of View"
@@ -13,7 +13,7 @@
 						:tooltip="{ text: 'The vertical field of view, from bottom to top of view' }"
 					>
 						<InputNumber
-							v-model="threeStore.activeCamera.fov"
+							v-model="cameraStore.activeCamera.fov"
 							:format-options="{
 								style: 'unit',
 								unitDisplay: 'narrow',
@@ -36,7 +36,7 @@
 					</InputField>
 				</template>
 				<InputField label="Zoom" input-width="150px">
-					<InputNumber v-model="threeStore.activeCamera.zoom" :step="0.01" />
+					<InputNumber v-model="cameraStore.activeCamera.zoom" :step="0.01" />
 				</InputField>
 				<InputField
 					label="Clip Start"
@@ -44,27 +44,27 @@
 					class="mt-1"
 					:tooltip="{ text: 'Camera near clipping distance' }"
 				>
-					<InputNumber v-model="threeStore.activeCamera.near" />
+					<InputNumber v-model="cameraStore.activeCamera.near" />
 				</InputField>
 				<InputField
 					label="End"
 					input-width="150px"
 					:tooltip="{ text: 'Camera far clipping distance' }"
 				>
-					<InputNumber v-model="threeStore.activeCamera.far" />
+					<InputNumber v-model="cameraStore.activeCamera.far" />
 				</InputField>
 			</div>
 		</MxAccordionItem>
 		<MxAccordionItem label="Transform" :item="{ value: 'transform' }">
 			<div class="p-1 flex flex-col gap-0.5 items-end">
 				<InputField label="Position X" input-width="150px">
-					<InputNumber v-model="threeStore.activeCamera.position.x" />
+					<InputNumber v-model="cameraStore.activeCamera.position.x" />
 				</InputField>
 				<InputField label="Y" input-width="150px">
-					<InputNumber v-model="threeStore.activeCamera.position.y" />
+					<InputNumber v-model="cameraStore.activeCamera.position.y" />
 				</InputField>
 				<InputField label="Z" input-width="150px">
-					<InputNumber v-model="threeStore.activeCamera.position.z" />
+					<InputNumber v-model="cameraStore.activeCamera.position.z" />
 				</InputField>
 			</div>
 		</MxAccordionItem>
@@ -72,36 +72,26 @@
 </template>
 
 <script lang="ts" setup>
-import { useThreeStore } from '@/store/three'
+import { useCameraStore } from '@/store/camera'
 import THREE from '@/three'
 import { computed, ref } from 'vue'
 
-const threeStore = useThreeStore()
-
-const cameraType = computed<string>({
-	set(val) {
-		if (val === cameraType.value) return
-		threeStore.switchCamera()
-	},
-	get() {
-		return threeStore.activeCamera.type
-	}
-})
+const cameraStore = useCameraStore()
 
 function isPerspectiveCamera(
 	camera: THREE.PerspectiveCamera | THREE.OrthographicCamera
 ): camera is THREE.PerspectiveCamera {
-	return threeStore.activeCamera.type === 'PerspectiveCamera'
+	return cameraStore.activeCamera.type === 'PerspectiveCamera'
 }
 
 const TYPE_OPTIONS = [
 	{
 		label: 'Perspective',
-		value: 'PerspectiveCamera'
+		value: 'perspective'
 	},
 	{
 		label: 'Orthographic',
-		value: 'OrthographicCamera'
+		value: 'orthographic'
 	}
 ] as const
 
@@ -109,13 +99,13 @@ const currentLensUnit = ref<(typeof LENS_UNIT_OPTIONS)[number]['value']>('fov')
 
 const focalLength = computed<number>({
 	set(val) {
-		if ('getFocalLength' in threeStore.activeCamera) {
-			return threeStore.activeCamera.setFocalLength(val)
+		if ('getFocalLength' in cameraStore.activeCamera) {
+			return cameraStore.activeCamera.setFocalLength(val)
 		}
 	},
 	get() {
-		if ('getFocalLength' in threeStore.activeCamera) {
-			return threeStore.activeCamera.getFocalLength()
+		if ('getFocalLength' in cameraStore.activeCamera) {
+			return cameraStore.activeCamera.getFocalLength()
 		}
 		return 0
 	}